+++ src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const availableMaps = ['world-1', 'world-2', 'world-3', 'nuts-1', 'nuts-2', 'nuts-3'];